+++ src/main/java/com/elmakers/mine/bukkit/spell/UndoableSpell.java
@@ -67,6 +67,12 @@ public abstract class UndoableSpell extends TargetingSpell {
 		getUndoList().addRemoved(entity);
 	}
 	
+	public void watch(Entity entity)
+	{
+		if (entity == null) return;
+		getUndoList().watch(entity);
+	}
+	
 	public UndoList getUndoList()
 	{
 		if (modifiedBlocks == null) modifiedBlocks = new UndoList(controller.getPlugin());