+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
  - Add "*1" format for multiplying directions.
  - Add MaterialBrush
- - Add UndoQueue and BlockList
+ - Add UndoQueue and BlockList, expand undo system to handle entities and falling blocks
+ - Add raw explosion method for assigning a source entity
 
 ## 1.0